Move FileTreeElement and RelativePath to base-services
